requirements.txt
pandas==2.0.3
websocket-client==1.6.1
requests==2.31.0
